refactor to more base classes
