+++ src/main/java/com/github/hcsp/inheritance/Animal.java
@@ -1,0 +1,16 @@
+package com.github.hcsp.inheritance;
+
+/**
+ * create by crashLab on 2020/06/16.
+ **/
+public class Animal {
+    private String name;
+
+    public Animal(String name) {
+        this.name = name;
+    }
+
+    public void sayMyName() {
+        System.out.println("我的名字是" + name);
+    }
+}
